Getting Updated UI (#25)
* Debugged views command (#22)

* Updating the UI

* Removed unnessasary imports

* Removed unessassary files
